Implemented the rhipo function 'getKernelArgumentMetaData' so that calling it within the hipRegisterFatBinary function in Sibir will get the binary data of an ELF .note section and store it as a JSON.
As of now, internal.cc will print out the names of the kernels as well as their vgpr_counts.

The CMakeLists.txt file had to be changed to include the elf.h and elf.cc files taken from Rhipo. We'll need to remember to update the CMakLists if more header files get added to this project.

One issue: the panic() function causes a symbol lookup error, I have no idea how to fix this but it is not very important.