Created getTextSection function. Maybe we could make a generic getSection Function?
Currently, the getTextSection function returns an elfio::Section pointer by calling the elfio::File.GetSectionByName function.
If we wanted to make a general getSection function, all we'd have to do is make the input of GetSectionByName an input.
So, it would be something like: `getSection(elfio::File *elf, std::string sectionname){ return elf->GetSectionByName(sectionname) }`
However, if we'll only be needing the text section, we shouldn't need to add or change anything else with this specific function.