Now we use the ELF FIle object to run the binary.
I changed the code so that we can take the char* from File.Blob() and pass it to a new wrapper's binary.
We do this by calculating the ELF file's size in the File.FromMem() function.
Then, we memcpy File.Blob into another char* of the same size as the ELF file.
This put into the binary of a new fat binary wrapper, and passed to exectuion.