Updated instprinter.h for disassembler library
Now a bunch of the bugs that were fixed in other branches are back,
so there is reworking that needs to be done. However, the vast majority
of instructions are now working properly, and for the ones that are not
working properly, the instruction name is disassembled correctly, but
the operands are wrong.

Do note that not all the examples have been retested.