Refactored assembler.cpp so that the program can take a single string assembly instruction as an argument, as opposed to multiple character pointers. I ended up changing input_file.cpp around a little while experimenting. Soon, the main assembler.cpp file will be ready to accept specially formatted .CSV files as arguments.
Eventually, I think the goal is to be able to parse a .S file, though I will have to check.
I also need to think about how the code will expand if/when I need to implement more instructions.

More immediate TO-DO: create an assembler class, like the disassembler class.
After that is done, incorporate the .CSV file parsing.