Changed registers of instruction STRINGS for inst linked list
Still need to assemble the edited instructions!!!!!

I need to fix the problems that occur when you try to compile the
assembler and disassembler together in one file. There are conflicts
because they call functions of the same name, from the same files,
but the files are copied into different places.