Modified Assembler to only be able to assemble SOP2, SOP1, and SOPP instructions.
This is because we only need this library to assemble our trampoline code, which
will only ever contain select instructions from those 3 formats. If we ever need
to add more, the assembler should now be modular enough to easily add support f-
or them.

Also, I added editGPR functions for SSRC0 and 1 as well as DST. This funciton w-
ill be used when offsetting regiter numbers.