Fixed issues with generated trampoline instruction, also got started with generating binary from instlist.
Corrected the SIMM16 values for the branch instructions
used to jump to and from the trampoline. Also, fixed the
issue where set and get pc regs inside the instrumentation
function were being offset when they were not supposed to.

WIP: created assembler function that generates a Blob from
an std::vector<std::shared_ptr<Inst>> -- need to verify if
it works.
  - this function was originally in bitops and was moved
    to the assembler.

    modified:   src_/instrumentation/copy_text/main.cpp
    modified:   src_/lib/src/assembler.cpp
    modified:   src_/lib/src/assembler.h
    modified:   src_/lib/src/bitops.cpp
    modified:   src_/lib/src/bitops.h